Remove int cast in texture upload

+++ MiniMikuDanceMaui/Rendering/PmxRenderer.cs
@@ -967,7 +967,7 @@ void main(){
                     GL.TexImage2D(
                         All.Texture2D,
                         0,
-                        (int)All.Rgba,
+                        All.Rgba,
                         sm.TextureWidth,
                         sm.TextureHeight,
                         0,
